Logic in radial_least_square_check()
The logic here is more consistent with the rest of the code if we use
the original `velbeam_arr` values rather than overwriting "unprocessed"
values with nans.

Furthermore, if the check is "good" (decision == 1) we should not update
final_vel or flag_vel -- flag_vel is already "good" or "unfolded" --
writing "good" to it loses information.